completion: make it a verb and factor it out into a separate file
