+++ mkosi/__init__.py
@@ -27,6 +27,7 @@ from typing import Optional, Union, cast
 
 from mkosi.archive import can_extract_tar, extract_tar, make_cpio, make_tar
 from mkosi.burn import run_burn
+from mkosi.completion import print_completion
 from mkosi.config import (
     PACKAGE_GLOBS,
     Args,
@@ -43,15 +44,10 @@ from mkosi.config import (
     Network,
     OutputFormat,
     SecureBootSignTool,
-    ShellCompletion,
     ShimBootloader,
     Verb,
     Vmm,
     __version__,
-    collect_completion_arguments,
-    finalize_completion_bash,
-    finalize_completion_fish,
-    finalize_completion_zsh,
     format_bytes,
     parse_config,
     summary,
@@ -4420,16 +4416,6 @@ def show_docs(args: Args, *, resources: Path) -> None:
                 raise e
 
 
-def print_completion(args: Args, *, resources: Path) -> None:
-    completion_args = collect_completion_arguments()
-    if args.shell_completion == ShellCompletion.bash:
-        print(finalize_completion_bash(completion_args, resources))
-    elif args.shell_completion == ShellCompletion.fish:
-        print(finalize_completion_fish(completion_args, resources))
-    elif args.shell_completion == ShellCompletion.zsh:
-        print(finalize_completion_zsh(completion_args, resources))
-
-
 def expand_specifier(s: str) -> str:
     return s.replace("%u", INVOKING_USER.name())
 
@@ -4770,7 +4756,7 @@ def run_verb(args: Args, images: Sequence[Config], *, resources: Path) -> None:
     if args.verb.needs_root() and os.getuid() != 0:
         die(f"Must be root to run the {args.verb} command")
 
-    if args.shell_completion:
+    if args.verb == Verb.completion:
         return print_completion(args, resources=resources)
 
     if args.verb == Verb.documentation:
