+++ tests/test_json.py
@@ -27,7 +27,6 @@ from mkosi.config import (
     QemuFirmware,
     QemuVsockCID,
     SecureBootSignTool,
-    ShellCompletion,
     ShimBootloader,
     Verb,
     Vmm,
@@ -56,7 +55,6 @@ def test_args(path: Optional[Path]) -> None:
             "GenkeyValidDays": "100",
             "Json": false,
             "Pager": true,
-            "ShellCompletion": "none",
             "Verb": "build"
         }}
         """
@@ -75,7 +73,6 @@ def test_args(path: Optional[Path]) -> None:
         genkey_valid_days = "100",
         json = False,
         pager = True,
-        shell_completion = ShellCompletion.none,
         verb = Verb.build,
     )
